Added documentations
