+++ composeChili/src/main/java/com/design/composechili/components/snackbar/BaseSnackBarParams.kt
@@ -1,0 +1,32 @@
+package com.design.composechili.components.snackbar
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.unit.Dp
+import com.design.composechili.theme.ChiliTextStyle
+import com.design.composechili.theme.ChiliTheme
+
+data class BaseSnackBarParams(
+    val containerColor: Color,
+    val cornersSize: Dp,
+    val textColor: Color,
+    val actionTextColor: Color,
+    val textStyle: TextStyle,
+) {
+    companion object {
+        val Default
+            @Composable
+            get() = BaseSnackBarParams(
+                containerColor = ChiliTheme.Colors.ChiliSnackbarBackground,
+                cornersSize = ChiliTheme.Attribute.ChiliSnackbarBackgroundCornerRadius,
+                textColor = ChiliTheme.Colors.ChiliSnackbarTextColor,
+                actionTextColor = ChiliTheme.Colors.ChiliComponentButtonTextColorActive,
+                textStyle = ChiliTextStyle.get(
+                    textSize = ChiliTheme.Attribute.ChiliTextDimensions.TextSizeH8,
+                    color = ChiliTheme.Colors.ChiliPrimaryTextColor,
+                )
+            )
+    }
+
+}
